MFC: ntp 4.1.0.
	See ChangeLog for all the many changes/bug fixes/new drivers.

	This code has been in CURRENT for months and several people use
	this version on 4.4-STABLE.

Approved by:	re

+++ contrib/ntp/libparse/parsesolaris.c
@@ -585,7 +585,7 @@ rdchar(
 			freeb(mmp);
 		}
 	}
-	return (unsigned)~0;
+	return (unsigned long)~0;
 }
 
 /*
